requirements.txt
numpy>=1.14.0
PyQt5>=5.12.0
pyqtgraph>=0.11.0rc0
pyserial>=3.0
requests>=2.20.0
scipy>=1.3.0
semver>=2.7.0
matplotlib>=3.0.0; sys_platform != 'win32'
matplotlib==3.2.2; sys_platform == 'win32'
importlib-resources>=1.1.0; python_version < '3.7'
dataclasses>=0.7; python_version < '3.7'
pywin32>=222; sys_platform == 'win32'
